ref(suite-common): move onPassphraseSubmit to common
