+++ suite-common/wallet-core/src/device/deviceThunks.ts
@@ -4,6 +4,7 @@ import TrezorConnect, {
     CardanoAddress,
     Address,
     Response as ConnectResponse,
+    UI,
 } from '@trezor/connect';
 import { TrezorDevice } from '@suite-common/suite-types';
 import { analytics, EventType } from '@trezor/suite-analytics';
@@ -519,5 +520,35 @@ export const confirmAddressOnDeviceThunk = createThunk(
                 } as const;
         }
         return response;
+    },
+);
+
+export const onPassphraseSubmit = createThunk(
+    `${MODULE_PREFIX}/onPassphraseSubmit`,
+    (
+        { value, passphraseOnDevice }: { value: string; passphraseOnDevice: boolean },
+        { dispatch, getState },
+    ) => {
+        const device = selectDevice(getState());
+        if (!device) return;
+
+        if (!device.state) {
+            dispatch(
+                deviceActions.updatePassphraseMode({
+                    device,
+                    hidden: passphraseOnDevice || !!value,
+                    alwaysOnDevice: passphraseOnDevice,
+                }),
+            );
+        }
+
+        TrezorConnect.uiResponse({
+            type: UI.RECEIVE_PASSPHRASE,
+            payload: {
+                value,
+                save: true,
+                passphraseOnDevice,
+            },
+        });
     },
 );
